Add window creation and property rotation feature
Updated `Program.cs` to implement a new feature for creating a window and rotating its properties using the `Xcsb` library. Added a new profile "RotatingValues" in `launchSettings.json` for easier execution. Created `RotatingValues.csproj` with necessary settings and a reference to `Xcsb.csproj`. Modified `PropertiesLength` in `RotatePropertiesType` struct in `Request.cs` to use `ushort` instead of `int`. Updated `Xcsb.sln` to include the new project and its build configurations for Debug and Release modes.

+++ Src/Models/Requests/Request.cs
@@ -154,7 +154,7 @@ internal readonly struct RotatePropertiesType(uint window, int propertiesLength,
     private readonly byte _pad0;
     public readonly ushort Length = (ushort)(3 + propertiesLength);
     public readonly uint Window = window;
-    public readonly int PropertiesLength = propertiesLength;
+    public readonly ushort PropertiesLength = (ushort)propertiesLength;
     public readonly ushort Delta = delta;
 }
